// onsite-program/submissions/week-4/day-4/day4-task2/Barakat-Yetunde-Oyewole/SchoolManagement/typechain-types/StudentManagementSystem.ts
export declare namespace StudentManagementSystem {
  export type StudentStruct = {
    name: string;
    age: BigNumberish;
    course: string;
    yearOfStudy: BigNumberish;
    status: BigNumberish;
    creator: AddressLike;
  };

  export type StudentStructOutput = [
    name: string,
    age: bigint,
    course: string,
    yearOfStudy: bigint,
    status: bigint,
    creator: string
  ] & {
    name: string;
    age: bigint;
    course: string;
    yearOfStudy: bigint;
    status: bigint;
    creator: string;
  };
}

export interface StudentManagementSystemInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "deleteStudent"
      | "getAllStudentAddresses"
      | "getAllStudents"
      | "getStudent"
      | "registerStudent"
      | "updateStudentInfo"
      | "updateStudentStatus"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "deleteStudent",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getAllStudentAddresses",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getAllStudents",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getStudent",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "registerStudent",
    values: [AddressLike, string, BigNumberish, string, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "updateStudentInfo",
    values: [AddressLike, string, BigNumberish, string, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "updateStudentStatus",
    values: [AddressLike, BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "deleteStudent",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getAllStudentAddresses",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getAllStudents",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "getStudent", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "registerStudent",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "updateStudentInfo",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "updateStudentStatus",
    data: BytesLike
  ): Result;
}

export interface StudentManagementSystem extends BaseContract {
  connect(runner?: ContractRunner | null): StudentManagementSystem;
  waitForDeployment(): Promise<this>;

  interface: StudentManagementSystemInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  deleteStudent: TypedContractMethod<
    [_studentAddress: AddressLike],
    [void],
    "nonpayable"
  >;

  getAllStudentAddresses: TypedContractMethod<[], [string[]], "view">;

  getAllStudents: TypedContractMethod<
    [],
    [StudentManagementSystem.StudentStructOutput[]],
    "view"
  >;

  getStudent: TypedContractMethod<
    [_studentAddress: AddressLike],
    [StudentManagementSystem.StudentStructOutput],
    "view"
  >;

  registerStudent: TypedContractMethod<
    [
      _studentAddress: AddressLike,
      _name: string,
      _age: BigNumberish,
      _course: string,
      _yearOfStudy: BigNumberish
    ],
    [void],
    "nonpayable"
  >;

  updateStudentInfo: TypedContractMethod<
    [
      _studentAddress: AddressLike,
      _name: string,
      _age: BigNumberish,
      _course: string,
      _yearOfStudy: BigNumberish
    ],
    [void],
    "nonpayable"
  >;

  updateStudentStatus: TypedContractMethod<
    [_studentAddress: AddressLike, _status: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "deleteStudent"
  ): TypedContractMethod<[_studentAddress: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "getAllStudentAddresses"
  ): TypedContractMethod<[], [string[]], "view">;
  getFunction(
    nameOrSignature: "getAllStudents"
  ): TypedContractMethod<
    [],
    [StudentManagementSystem.StudentStructOutput[]],
    "view"
  >;
  getFunction(
    nameOrSignature: "getStudent"
  ): TypedContractMethod<
    [_studentAddress: AddressLike],
    [StudentManagementSystem.StudentStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "registerStudent"
  ): TypedContractMethod<
    [
      _studentAddress: AddressLike,
      _name: string,
      _age: BigNumberish,
      _course: string,
      _yearOfStudy: BigNumberish
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "updateStudentInfo"
  ): TypedContractMethod<
    [
      _studentAddress: AddressLike,
      _name: string,
      _age: BigNumberish,
      _course: string,
      _yearOfStudy: BigNumberish
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "updateStudentStatus"
  ): TypedContractMethod<
    [_studentAddress: AddressLike, _status: BigNumberish],
    [void],
    "nonpayable"
  >;

  filters: {};
}